add access to R's `identical()` function and skip testing equality for lists
List equality is not implemented in R for the == operator. It throws an
error.

I thought that I could use R's `identical(x, y)` function to fix that,
but while it worked in R like:
    > x <- list(1, 2, 3)
    > y <- list(1, 2, 3)
    > identical(x, y)
    [1] TRUE
calling `identical` on the lists via the C API did not work. It also did
not work for other data built on `VECSXP`s such as `data.frame`s.