Test matrix construction more thoroughly.
We need to make sure that the first dimension in R and Perl refer to the
same parts of memory. This is why we test the slicing syntax as well and
the specific indices inside a matrix.

The index [1,2] in R should have the same value as the position [0,1] in
Perl.